third_party/WebKit/LayoutTests/accessibility/aria-hidden-children-not-in-text-from-content-expected.txt

This tests that aria-hidden elements within an element are not included in accessible name from content

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".



PASS checkbox1.title.trim() is 'AXTitle:'
PASS checkbox2.title.trim() is 'AXTitle: Non-hidden descendant content'
